chore: rename file removing typo
