+++ backend/src/common/constants.ts
@@ -17,3 +17,8 @@ export const REFRESH_TOKEN_COOKIE_OPTIONS = {
 export const EMAIL_AUTH_EXPIRE_TIME = 5 * SECONDS_PER_MINUTE
 
 export const PUBLICIZING_REQUEST_EXPIRE_TIME = 7 * SECONDS_PER_DAY
+
+export const PUBLIC_GROUP_ID = 1
+export const PAGINATION_MAX_LIMIT = 10
+export const PAGINATION_MIN_LIMIT = 0
+export const PAGINATION_MIN_OFFSET = 0
