--- backend/src/common/contstants.ts
@@ -1,4 +1,0 @@
-export const PUBLIC_GROUP_ID = 1
-export const PAGINATION_MAX_LIMIT = 10
-export const PAGINATION_MIN_LIMIT = 0
-export const PAGINATION_MIN_OFFSET = 0
