+++ backend/src/problem/problem.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common'
 import { ContestProblem, Problem, WorkbookProblem } from '@prisma/client'
-import { PUBLIC_GROUP_ID } from 'src/common/contstants'
+import { PUBLIC_GROUP_ID } from 'src/common/constants'
 import { PaginationDto } from 'src/common/dto/pagination.dto'
 import { EntityNotExistException } from 'src/common/exception/business.exception'
 import { PrismaService } from 'src/prisma/prisma.service'
