+++ backend/src/problem/problem.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common'
 import { plainToInstance } from 'class-transformer'
-import { PUBLIC_GROUP_ID } from 'src/common/contstants'
+import { PUBLIC_GROUP_ID } from 'src/common/constants'
 import { PaginationDto } from 'src/common/dto/pagination.dto'
 import { EntityNotExistException } from 'src/common/exception/business.exception'
 import { ContestService } from 'src/contest/contest.service'
